+++ CHANGELOG.md
@@ -10,13 +10,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- This CHANGELOG
 - Outposts included in `objectives_count`
 - Update for Antistasi Ultimate v11.6.0 (already included new map, but text had old version number)
+- Site text: point out table is sortable, link to CHANGELOG
 
 ### Fixed
 
 - Airports weren't included in `objectives_count`
 - Text implied bases weren't included in `objectives_count`
+- Site text: Heading levels
 
 
 ## [0.2.0] - 2025-04-09